+++ .github/workflows/python-integration-tests.yml
@@ -33,14 +33,23 @@ jobs:
       uses: actions/setup-python@v4
       with:
         python-version: ${{ matrix.python-version }}
-    - name: Install dependencies
+    - name: Install dependencies wtih hnswlib native disabled
+      if: matrix.os == 'macos-latest' && matrix.python-version == '3.11'
       run: |
-        python -m pip install --upgrade pip
+        export HNSWLIB_NO_NATIVE=1
+        python -m pip install --upgrade pip setuptools wheel
+        python -m pip install poetry pytest
+        cd python && poetry install --with hugging_face --with chromadb
+    - name: Install dependencies wtih hnswlib native enabled
+      if: matrix.os != 'macos-latest' || matrix.python-version != '3.11'
+      run: |
+        python -m pip install --upgrade pip setuptools wheel
         python -m pip install poetry pytest
         cd python && poetry install --with hugging_face --with chromadb
     - name: Run Integration Tests
       shell: bash
       env: # Set Azure credentials secret as an input
+        HNSWLIB_NO_NATIVE: 1
         Python_Integration_Tests: Python_Integration_Tests
         AzureOpenAI__Label: azure-text-davinci-003
         AzureOpenAIEmbedding__Label: azure-text-embedding-ada-002
