+++ python/tests/integration/completions/test_hf_local_text_summarization.py
@@ -1,28 +1,129 @@
 # Copyright (c) Microsoft. All rights reserved.
 
-import asyncio
-
-import e2e_text_completion
 import pytest
 
 import semantic_kernel as sk
-import semantic_kernel.connectors.ai.hugging_face as sk_hf
 
 
 @pytest.mark.asyncio
-async def test_hf_local_summarization_service_with_skills():
-    kernel = sk.Kernel()
+async def test_summarize_input_str(setup_summarize_function):
+    (
+        kernel,
+        summarize_function,
+        text_to_summarize,
+        additional_text,
+    ) = setup_summarize_function
 
-    # Configure LLM service
-    kernel.add_text_completion_service(
-        "facebook/bart-large-cnn",
-        sk_hf.HuggingFaceTextCompletion(
-            "facebook/bart-large-cnn", task="summarization"
-        ),
+    # Summarize input string and print
+    summary = await kernel.run_async(summarize_function, input_str=text_to_summarize)
+
+    output = str(summary).strip()
+    print(f"Summary using input string: '{output}'")
+    assert len(output) > 0
+
+
+@pytest.mark.asyncio
+async def test_summarize_input_vars(setup_summarize_function):
+    (
+        kernel,
+        summarize_function,
+        text_to_summarize,
+        additional_text,
+    ) = setup_summarize_function
+
+    # Summarize input as context variable and print
+    context_vars = sk.ContextVariables(text_to_summarize)
+    summary = await kernel.run_async(summarize_function, input_vars=context_vars)
+
+    output = str(summary).strip()
+    print(f"Summary using context variables: '{output}'")
+    assert len(output) > 0
+
+
+@pytest.mark.asyncio
+async def test_summarize_input_context(setup_summarize_function):
+    (
+        kernel,
+        summarize_function,
+        text_to_summarize,
+        additional_text,
+    ) = setup_summarize_function
+
+    # Summarize input context and print
+    context = kernel.create_new_context()
+    context["input"] = text_to_summarize
+    summary = await kernel.run_async(summarize_function, input_context=context)
+
+    output = str(summary).strip()
+    print(f"Summary using input context: '{output}'")
+    assert len(output) > 0
+
+
+@pytest.mark.asyncio
+async def test_summarize_input_context_with_vars(setup_summarize_function):
+    (
+        kernel,
+        summarize_function,
+        text_to_summarize,
+        additional_text,
+    ) = setup_summarize_function
+
+    # Summarize input context with additional variables and print
+    context = kernel.create_new_context()
+    context["input"] = text_to_summarize
+    context_vars = sk.ContextVariables(additional_text)
+    summary = await kernel.run_async(
+        summarize_function, input_context=context, input_vars=context_vars
     )
 
-    await e2e_text_completion.simple_summarization(kernel)
+    output = str(summary).strip()
+    print(f"Summary using context and additional variables: '{output}'")
+    assert len(output) > 0
 
 
-if __name__ == "__main__":
-    asyncio.run(test_hf_local_summarization_service_with_skills())
+@pytest.mark.asyncio
+async def test_summarize_input_context_with_str(setup_summarize_function):
+    (
+        kernel,
+        summarize_function,
+        text_to_summarize,
+        additional_text,
+    ) = setup_summarize_function
+
+    # Summarize input context with additional input string and print
+    context = kernel.create_new_context()
+    context["input"] = text_to_summarize
+    summary = await kernel.run_async(
+        summarize_function, input_context=context, input_str=additional_text
+    )
+
+    output = str(summary).strip()
+    print(f"Summary using context and additional string: '{output}'")
+    assert len(output) > 0
+
+
+@pytest.mark.asyncio
+async def test_summarize_input_context_with_vars_and_str(setup_summarize_function):
+    (
+        kernel,
+        summarize_function,
+        text_to_summarize,
+        additional_text,
+    ) = setup_summarize_function
+
+    # Summarize input context with additional variables and string and print
+    context = kernel.create_new_context()
+    context["input"] = text_to_summarize
+    context_vars = sk.ContextVariables(variables={"input2": additional_text})
+    summary = await kernel.run_async(
+        summarize_function,
+        input_context=context,
+        input_vars=context_vars,
+        input_str="new text",
+    )
+
+    output = str(summary).strip()
+    print(
+        f"Summary using context, additional variables, and additional string: '{output}'"
+    )
+    assert len(output) > 0
