+++ python/tests/integration/completions/test_utils.py
@@ -1,0 +1,23 @@
+# Copyright (c) Microsoft. All rights reserved.
+
+import logging
+import time
+
+logging.basicConfig(level=logging.DEBUG)
+logger = logging.getLogger()
+
+
+async def retry(func, retries=20):
+    min_delay = 2
+    max_delay = 7
+    for i in range(retries):
+        try:
+            result = str(await func())
+            if "Error" in result:
+                raise ValueError(result)
+            return result
+        except Exception as e:
+            logger.error(f"Retry {i + 1}: {e}")
+            if i == retries - 1:  # Last retry
+                raise
+            time.sleep(max(min(i, max_delay), min_delay))
