+++ python/tests/integration/embeddings/test_oai_embedding_service.py
@@ -1,9 +1,5 @@
 # Copyright (c) Microsoft. All rights reserved.
 
-import asyncio
-import os
-
-import e2e_memories
 import pytest
 
 import semantic_kernel as sk
@@ -11,15 +7,10 @@ import semantic_kernel.connectors.ai.open_ai as sk_oai
 
 
 @pytest.mark.asyncio
-async def test_oai_embedding_service_with_memories():
-    kernel = sk.Kernel()
+async def test_oai_embedding_service(create_kernel, get_oai_config):
+    kernel = create_kernel
 
-    if "Python_Integration_Tests" in os.environ:
-        api_key = os.environ["OpenAI__ApiKey"]
-        org_id = None
-    else:
-        # Load credentials from .env file
-        api_key, org_id = sk.openai_settings_from_dot_env()
+    api_key, org_id = get_oai_config
 
     kernel.add_text_embedding_generation_service(
         "oai-ada",
@@ -27,8 +18,12 @@ async def test_oai_embedding_service_with_memories():
     )
     kernel.register_memory_store(memory_store=sk.memory.VolatileMemoryStore())
 
-    await e2e_memories.simple_memory_test(kernel)
-
-
-if __name__ == "__main__":
-    asyncio.run(test_oai_embedding_service_with_memories())
+    await kernel.memory.save_information_async(
+        "test", id="info1", text="this is a test"
+    )
+    await kernel.memory.save_reference_async(
+        "test",
+        external_id="info1",
+        text="this is a test",
+        external_source_name="external source",
+    )
