+++ python/tests/unit/core_skills/test_file_io_skill.py
@@ -3,8 +3,8 @@ import tempfile
 
 import pytest
 
-from semantic_kernel.core_skills.file_io_skill import FileIOSkill
 from semantic_kernel import Kernel
+from semantic_kernel.core_skills.file_io_skill import FileIOSkill
 from semantic_kernel.orchestration.context_variables import ContextVariables
 from semantic_kernel.orchestration.sk_context import SKContext
